restructure redux utilities
